refactorying draggable
